refactor(policies): extract find-and-replace hook and popover
Move TipTap search/replace state, handlers, and popover UI into co-located PolicyEditor modules to shrink the page.
