+++ Clients/src/presentation/pages/PolicyDashboard/PolicyEditorPage.tsx
@@ -78,14 +78,11 @@ import {
   Check,
   Pencil,
   Palette,
-  ChevronDown as ChevronDownIcon,
-  ChevronUp as ChevronUpIcon,
   Search,
   Upload,
 } from "lucide-react";
 
 import Select from "../../components/Inputs/Select";
-import Field from "../../components/Inputs/Field";
 import { CustomizableButton } from "../../components/button/customizable-button";
 import { HistorySidebar } from "../../components/Common/HistorySidebar";
 import CustomFieldsSection, {
@@ -113,12 +110,11 @@ import { store } from "../../../application/redux/store";
 import { PageBreadcrumbs } from "../../components/breadcrumbs/PageBreadcrumbs";
 import { AuthImageExtension } from "./PolicyEditor/AuthImage";
 import { normalizeSlateHtml } from "./PolicyEditor/normalizeSlateHtml";
-import {
-  createSearchHighlightExtension,
-  searchHighlightKey,
-} from "./PolicyEditor/searchHighlightExtension";
+import { createSearchHighlightExtension } from "./PolicyEditor/searchHighlightExtension";
 import { type ToolbarKey, defaultToolbarState } from "./PolicyEditor/toolbarTypes";
 import { policyEditorStyles } from "./PolicyEditor/editorStyles";
+import { usePolicyFindReplace } from "./PolicyEditor/usePolicyFindReplace";
+import { FindReplacePopover } from "./PolicyEditor/FindReplacePopover";
 
 // ── Component ─────────────────────────────────────────────────────────
 export default function PolicyEditorPage() {
@@ -183,12 +179,6 @@ export default function PolicyEditorPage() {
 
   // Color picker state
   const [colorAnchorEl, setColorAnchorEl] = useState<HTMLElement | null>(null);
-
-  // Search & replace state
-  const [searchAnchorEl, setSearchAnchorEl] = useState<HTMLElement | null>(null);
-  const [searchText, setSearchText] = useState("");
-  const [replaceText, setReplaceText] = useState("");
-  const [searchMatchCount, setSearchMatchCount] = useState(0);
 
   const validators = useMemo(
     () => ({
@@ -583,171 +573,20 @@ export default function PolicyEditorPage() {
     "#9333ea",
   ];
 
-  // ── Search & replace ───────────────────────────────────────────
-  const countMatches = useCallback(() => {
-    if (!editor || !searchText) {
-      setSearchMatchCount(0);
-      return;
-    }
-    const searchLower = searchText.toLowerCase();
-    let count = 0;
-    editor.state.doc.descendants((node) => {
-      if (!node.isText || !node.text) return;
-      const text = node.text.toLowerCase();
-      let idx = text.indexOf(searchLower);
-      while (idx !== -1) {
-        count++;
-        idx = text.indexOf(searchLower, idx + searchText.length);
-      }
-    });
-    setSearchMatchCount(count);
-  }, [editor, searchText]);
-
-  // Count matches, update decorations, and auto-find first match when searchText changes
-  useEffect(() => {
-    countMatches();
-    if (!editor) return;
-    // Update search highlight decorations
-    const tr = editor.state.tr.setMeta(searchHighlightKey, searchText || "");
-    editor.view.dispatch(tr);
-    // Auto-jump to first match
-    if (searchText) {
-      const searchLower = searchText.toLowerCase();
-      let found = false;
-      editor.state.doc.descendants((node, pos) => {
-        if (found || !node.isText || !node.text) return;
-        const text = node.text.toLowerCase();
-        const idx = text.indexOf(searchLower);
-        if (idx !== -1) {
-          const from = pos + idx;
-          const to = from + searchText.length;
-          editor.chain().setTextSelection({ from, to }).scrollIntoView().run();
-          found = true;
-        }
-      });
-    }
-  }, [countMatches, editor, searchText]);
-
-  const handleSearchNext = useCallback(() => {
-    if (!editor || !searchText) return;
-    const { doc, selection } = editor.state;
-    const searchLower = searchText.toLowerCase();
-    const startFrom = selection.to;
-    let found = false;
-
-    doc.descendants((node, pos) => {
-      if (found || !node.isText || !node.text) return;
-      const text = node.text.toLowerCase();
-      const idx =
-        pos >= startFrom ? text.indexOf(searchLower) : text.indexOf(searchLower, startFrom - pos);
-      if (idx !== -1 && pos + idx >= startFrom) {
-        const from = pos + idx;
-        const to = from + searchText.length;
-        editor.chain().setTextSelection({ from, to }).scrollIntoView().run();
-        found = true;
-      }
-    });
-
-    // Wrap around if not found after cursor
-    if (!found) {
-      doc.descendants((node, pos) => {
-        if (found || !node.isText || !node.text) return;
-        const text = node.text.toLowerCase();
-        const idx = text.indexOf(searchLower);
-        if (idx !== -1) {
-          const from = pos + idx;
-          const to = from + searchText.length;
-          editor.chain().setTextSelection({ from, to }).scrollIntoView().run();
-          found = true;
-        }
-      });
-    }
-  }, [editor, searchText]);
-
-  const handleSearchPrev = useCallback(() => {
-    if (!editor || !searchText) return;
-    const { doc, selection } = editor.state;
-    const searchLower = searchText.toLowerCase();
-    const startBefore = selection.from;
-    let lastMatch: { from: number; to: number } | null = null;
-
-    // Find the last match before cursor
-    doc.descendants((node, pos) => {
-      if (!node.isText || !node.text) return;
-      const text = node.text.toLowerCase();
-      let idx = text.indexOf(searchLower);
-      while (idx !== -1) {
-        const from = pos + idx;
-        if (from < startBefore) {
-          lastMatch = { from, to: from + searchText.length };
-        }
-        idx = text.indexOf(searchLower, idx + searchText.length);
-      }
-    });
-
-    if (lastMatch) {
-      editor.chain().setTextSelection(lastMatch).scrollIntoView().run();
-    } else {
-      // Wrap around: find last match in the whole doc
-      doc.descendants((node, pos) => {
-        if (!node.isText || !node.text) return;
-        const text = node.text.toLowerCase();
-        let idx = text.indexOf(searchLower);
-        while (idx !== -1) {
-          const from = pos + idx;
-          lastMatch = { from, to: from + searchText.length };
-          idx = text.indexOf(searchLower, idx + searchText.length);
-        }
-      });
-      if (lastMatch) {
-        editor.chain().setTextSelection(lastMatch).scrollIntoView().run();
-      }
-    }
-  }, [editor, searchText]);
-
-  const handleReplaceCurrent = useCallback(() => {
-    if (!editor || !searchText) return;
-    const { from, to } = editor.state.selection;
-    const selectedText = editor.state.doc.textBetween(from, to).toLowerCase();
-    if (selectedText === searchText.toLowerCase()) {
-      editor.chain().deleteSelection().insertContent(replaceText).run();
-      countMatches();
-      // Find next match after replacement
-      handleSearchNext();
-    } else {
-      // No current selection matching — find next first
-      handleSearchNext();
-    }
-  }, [editor, searchText, replaceText, countMatches, handleSearchNext]);
-
-  const handleReplaceAll = useCallback(() => {
-    if (!editor || !searchText) return;
-    const { doc, tr } = editor.state;
-    const searchLower = searchText.toLowerCase();
-    let replaced = 0;
-
-    doc.descendants((node, pos) => {
-      if (!node.isText || !node.text) return;
-      const text = node.text.toLowerCase();
-      let idx = text.indexOf(searchLower);
-      while (idx !== -1) {
-        const from = pos + idx;
-        const to = from + searchText.length;
-        tr.replaceWith(
-          from + replaced * (replaceText.length - searchText.length),
-          to + replaced * (replaceText.length - searchText.length),
-          editor.state.schema.text(replaceText),
-        );
-        replaced++;
-        idx = text.indexOf(searchLower, idx + searchText.length);
-      }
-    });
-
-    if (replaced > 0) {
-      editor.view.dispatch(tr);
-      countMatches();
-    }
-  }, [editor, searchText, replaceText, countMatches]);
+  const {
+    searchAnchorEl,
+    openFindReplace,
+    closeFindReplace,
+    searchText,
+    setSearchText,
+    replaceText,
+    setReplaceText,
+    searchMatchCount,
+    handleSearchNext,
+    handleSearchPrev,
+    handleReplaceCurrent,
+    handleReplaceAll,
+  } = usePolicyFindReplace(editor);
 
   // ── Toolbar config ────────────────────────────────────────────────
   const toolbarConfig: Array<{
@@ -1671,7 +1510,7 @@ export default function PolicyEditorPage() {
                       return;
                     }
                     if (key === "search") {
-                      setSearchAnchorEl(e.currentTarget);
+                      openFindReplace(e.currentTarget);
                       return;
                     }
                     action?.();
@@ -1761,157 +1600,19 @@ export default function PolicyEditorPage() {
             </Box>
           </Popover>
 
-          {/* ── Find & replace popover (Google Docs style) ────────────── */}
-          <Popover
-            open={Boolean(searchAnchorEl)}
+          <FindReplacePopover
             anchorEl={searchAnchorEl}
-            onClose={() => {
-              setSearchAnchorEl(null);
-              setSearchText("");
-              setReplaceText("");
-            }}
-            anchorOrigin={{ vertical: "bottom", horizontal: "right" }}
-            transformOrigin={{ vertical: "top", horizontal: "right" }}
-            disableAutoFocus
-            disableEnforceFocus
-            slotProps={{
-              paper: {
-                sx: {
-                  width: 340,
-                  borderRadius: "4px",
-                  border: "1px solid #d0d5dd",
-                  boxShadow: "0 4px 16px rgba(0,0,0,0.12)",
-                  p: "16px",
-                },
-              },
-            }}
-          >
-            {/* Find row */}
-            <Stack gap="12px">
-              <Stack direction="row" gap="8px" alignItems="center">
-                <Box sx={{ flex: 1 }}>
-                  <Field
-                    id="search-find-input"
-                    placeholder="Find in document..."
-                    value={searchText}
-                    onChange={(e) => setSearchText(e.target.value)}
-                    onKeyDown={(e) => {
-                      if (e.key === "Enter") handleSearchNext();
-                    }}
-                    autoFocus
-                    sx={{
-                      "& .field-input": { height: 34 },
-                      "& input": { padding: "0 10px", fontSize: 13 },
-                    }}
-                  />
-                </Box>
-                <Tooltip title="Previous" arrow>
-                  <IconButton
-                    onClick={handleSearchPrev}
-                    disabled={searchMatchCount === 0}
-                    size="small"
-                    sx={{
-                      "padding": "6px",
-                      "borderRadius": "4px",
-                      "border": "1px solid #d0d5dd",
-                      "color": "text.secondary",
-                      "&:hover": { backgroundColor: "background.accent" },
-                      "&:disabled": { color: "border.dark", borderColor: "border.light" },
-                    }}
-                  >
-                    <ChevronUpIcon size={16} />
-                  </IconButton>
-                </Tooltip>
-                <Tooltip title="Next" arrow>
-                  <IconButton
-                    onClick={handleSearchNext}
-                    disabled={searchMatchCount === 0}
-                    size="small"
-                    sx={{
-                      "padding": "6px",
-                      "borderRadius": "4px",
-                      "border": "1px solid #d0d5dd",
-                      "color": "text.secondary",
-                      "&:hover": { backgroundColor: "background.accent" },
-                      "&:disabled": { color: "border.dark", borderColor: "border.light" },
-                    }}
-                  >
-                    <ChevronDownIcon size={16} />
-                  </IconButton>
-                </Tooltip>
-              </Stack>
-
-              {/* Match count */}
-              {searchText && (
-                <Typography sx={{ fontSize: 11, color: "text.muted", mt: "-4px" }}>
-                  {searchMatchCount === 0
-                    ? "No matches found"
-                    : `${searchMatchCount} match${searchMatchCount !== 1 ? "es" : ""} found`}
-                </Typography>
-              )}
-
-              {/* Replace row */}
-              <Stack direction="row" gap="8px" alignItems="center">
-                <Box sx={{ flex: 1 }}>
-                  <Field
-                    id="search-replace-input"
-                    placeholder="Replace with..."
-                    value={replaceText}
-                    onChange={(e) => setReplaceText(e.target.value)}
-                    onKeyDown={(e) => {
-                      if (e.key === "Enter") handleReplaceCurrent();
-                    }}
-                    sx={{
-                      "& .field-input": { height: 34 },
-                      "& input": { padding: "0 10px", fontSize: 13 },
-                    }}
-                  />
-                </Box>
-                <Tooltip title="Replace" arrow>
-                  <span>
-                    <CustomizableButton
-                      variant="outlined"
-                      text="Replace"
-                      onClick={handleReplaceCurrent}
-                      isDisabled={searchMatchCount === 0}
-                      sx={{
-                        "minWidth": "auto",
-                        "height": 34,
-                        "px": "10px",
-                        "fontSize": 12,
-                        "backgroundColor": "background.main",
-                        "border": "1px solid #d0d5dd",
-                        "color": "text.secondary",
-                        "whiteSpace": "nowrap",
-                        "&:hover": { backgroundColor: "background.accent" },
-                      }}
-                    />
-                  </span>
-                </Tooltip>
-              </Stack>
-
-              {/* Replace all */}
-              <Box sx={{ display: "flex", justifyContent: "flex-end" }}>
-                <CustomizableButton
-                  variant="outlined"
-                  text="Replace all"
-                  onClick={handleReplaceAll}
-                  isDisabled={searchMatchCount === 0}
-                  sx={{
-                    "minWidth": "auto",
-                    "height": 34,
-                    "px": "10px",
-                    "fontSize": 12,
-                    "backgroundColor": "background.main",
-                    "border": "1px solid #d0d5dd",
-                    "color": "text.secondary",
-                    "whiteSpace": "nowrap",
-                    "&:hover": { backgroundColor: "background.accent" },
-                  }}
-                />
-              </Box>
-            </Stack>
-          </Popover>
+            onClose={closeFindReplace}
+            searchText={searchText}
+            onSearchTextChange={setSearchText}
+            replaceText={replaceText}
+            onReplaceTextChange={setReplaceText}
+            searchMatchCount={searchMatchCount}
+            onSearchNext={handleSearchNext}
+            onSearchPrev={handleSearchPrev}
+            onReplaceCurrent={handleReplaceCurrent}
+            onReplaceAll={handleReplaceAll}
+          />
 
           {/* ── Editor + History sidebar ────────────────────────────── */}
           <Stack direction="row" sx={{ flex: 1, minHeight: 0, overflow: "hidden" }}>
